src/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
yoyo-migrations==8.2.0
sqlmodel==0.0.22
alembic==1.13.1